Move the new files into the package
