+++ imagematerials/buildings/constants.py
