+++ imagematerials/buildings/preprocessing/floorspace.py
@@ -1,25 +1,26 @@
 
-import pandas as pd
-
-import pandas as pd
-import numpy as np
-import os
 import math
-from pathlib import Path
-from matplotlib import pyplot as plt
-import xarray as xr
-from read_mym import read_mym_df
-from util import dataset_to_array
 
-from constants_BUMA import(
-    SCENARIO_SELECT, FILE_ADDITION,
-    REGIONS, REGIONS_RANGE,
-    START_YEAR, END_YEAR, HIST_YEAR, YEARS, YEAR_LIST_SVA,
-    INFLATION, 
-    FLAG_ALPHA, FLAG_EXPDEC, FLAG_NORMAL,
-    LOWCOMM,
+import numpy as np
+import pandas as pd
+import xarray as xr
+
+from imagematerials.buildings.constants import (
+    END_YEAR,
+    FLAG_ALPHA,
+    FLAG_EXPDEC,
     GOMPERTZ_EXPDEC,
+    HIST_YEAR,
+    INFLATION,
+    REGIONS,
+    REGIONS_RANGE,
+    START_YEAR,
+    YEAR_LIST_SVA,
+    YEARS,
 )
+from imagematerials.read_mym import read_mym_df
+from imagematerials.util import dataset_to_array
+
 far_start_year = 1721
 start_year = 1820
 end_year = 1970
