+++ imagematerials/buildings/preprocessing/population.py
@@ -1,10 +1,11 @@
-import pandas as pd
-import numpy as np
 from pathlib import Path
-import xarray as xr
-from util import dataset_to_array
 
-from constants_BUMA import SCENARIO_SELECT, YEARS
+import numpy as np
+import pandas as pd
+import xarray as xr
+
+from imagematerials.buildings.constants import SCENARIO_SELECT, YEARS
+from imagematerials.util import dataset_to_array
 
 far_start_year = 1721
 start_year = 1820
