+++ imagematerials/util.py
